[no-issue]: Extract permission checkers
